+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/PinnedSiteStorage.kt
@@ -75,4 +75,16 @@ class PinnedSiteStorage(context: Context) {
     suspend fun removePinnedSite(site: TopSite) = withContext(IO) {
         pinnedSiteDao.deletePinnedSite(site.toPinnedSite())
     }
+
+    /**
+     * Renames the given pinned site.
+     *
+     * @param site The pinned site.
+     * @param title The new title for the top site.
+     */
+    suspend fun renamePinnedSite(site: TopSite, title: String) = withContext(IO) {
+        val pinnedSite = site.toPinnedSite()
+        pinnedSite.title = title
+        pinnedSiteDao.updatePinnedSite(pinnedSite)
+    }
 }