+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/TopSitesUseCases.kt
@@ -37,11 +37,30 @@ class TopSitesUseCases(topSitesStorage: TopSitesStorage) {
         }
     }
 
+    /**
+     * Rename a top site use case.
+     */
+    class RenameTopSiteUseCase internal constructor(private val storage: TopSitesStorage) {
+        /**
+         * Renames the given [TopSite].
+         *
+         * @param topSite The top site.
+         * @param title The new title for the top site.
+         */
+        operator fun invoke(topSite: TopSite, title: String) {
+            storage.renameTopSite(topSite, title)
+        }
+    }
+
     val addPinnedSites: AddPinnedSiteUseCase by lazy {
         AddPinnedSiteUseCase(topSitesStorage)
     }
 
     val removeTopSites: RemoveTopSiteUseCase by lazy {
         RemoveTopSiteUseCase(topSitesStorage)
+    }
+
+    val renameTopSites: RenameTopSiteUseCase by lazy {
+        RenameTopSiteUseCase(topSitesStorage)
     }
 }