+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/db/PinnedSiteDao.kt
@@ -10,6 +10,7 @@ import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
+import androidx.room.Update
 
 /**
  * Internal DAO for accessing [PinnedSiteEntity] instances.
@@ -19,6 +20,10 @@ internal interface PinnedSiteDao {
     @WorkerThread
     @Insert
     fun insertPinnedSite(site: PinnedSiteEntity): Long
+
+    @WorkerThread
+    @Update
+    fun updatePinnedSite(site: PinnedSiteEntity)
 
     @WorkerThread
     @Delete