+++ mobile/android/android-components/components/feature/top-sites/src/test/java/mozilla/components/feature/top/sites/TopSitesUseCasesTest.kt
@@ -37,4 +37,16 @@ class TopSitesUseCasesTest {
 
         verify(topSitesStorage).removeTopSite(topSite)
     }
+
+    @Test
+    fun `RenameTopSiteUseCase`() = runBlocking {
+        val topSitesStorage: TopSitesStorage = mock()
+        val topSite: TopSite = mock()
+        val useCases = TopSitesUseCases(topSitesStorage)
+
+        val title = "New title"
+        useCases.renameTopSites(topSite, title)
+
+        verify(topSitesStorage).renameTopSite(topSite, title)
+    }
 }